+++ packages/form-plugin/src/actions.ts
@@ -82,3 +82,11 @@ export class SetFormEnabled {
   }
   constructor(public payload: string) {}
 }
+
+export class ResetForm {
+  static get type() {
+    // NOTE: Not necessary to declare the type in this way in your code. See https://github.com/ngxs/store/pull/644#issuecomment-436003138
+    return '[Forms] Reset Form';
+  }
+  constructor(public payload: { path: string; value?: any }) {}
+}