+++ packages/form-plugin/tests/form.plugin.spec.ts
@@ -1,21 +1,22 @@
 import { Component, Injectable } from '@angular/core';
-import { TestBed, fakeAsync, tick } from '@angular/core/testing';
-import { ReactiveFormsModule, FormGroup, FormControl, Validators } from '@angular/forms';
+import { fakeAsync, TestBed, tick } from '@angular/core/testing';
+import { FormControl, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
 import { By } from '@angular/platform-browser';
 
-import { NgxsModule, State, Store, Actions, ofActionDispatched, Selector } from '@ngxs/store';
+import { Actions, NgxsModule, ofActionDispatched, Selector, State, Store } from '@ngxs/store';
 
 import {
   NgxsFormPluginModule,
+  ResetForm,
+  SetFormDirty,
+  SetFormDisabled,
+  SetFormEnabled,
+  SetFormPristine,
   UpdateForm,
   UpdateFormDirty,
   UpdateFormErrors,
   UpdateFormStatus,
-  UpdateFormValue,
-  SetFormDirty,
-  SetFormDisabled,
-  SetFormEnabled,
-  SetFormPristine
+  UpdateFormValue
 } from '../';
 
 describe('NgxsFormPlugin', () => {
@@ -205,6 +206,52 @@ describe('NgxsFormPlugin', () => {
     expect(form.errors!.address).toEqual('address is too long');
     expect(form.model.name).toEqual('Lou Grant');
     expect(form.model.address).toEqual('waterloo, ontario');
+  });
+
+  it('should reset form', async () => {
+    // Arrange & act
+    const store = getStore();
+
+    await store
+      .dispatch([
+        new UpdateForm({
+          errors: {},
+          dirty: true,
+          status: 'INVALID',
+          value: { name: 'Ozan Turhan' },
+          path: 'student.studentForm'
+        }),
+        new ResetForm({ path: 'student.studentForm' })
+      ])
+      .toPromise();
+
+    const form = store.selectSnapshot(StudentState.getStudentForm);
+
+    // Assert
+    expect(form).toEqual({ model: {} });
+  });
+
+  it('should reset form with value', async () => {
+    // Arrange & act
+    const store = getStore();
+
+    await store
+      .dispatch([
+        new UpdateForm({
+          errors: {},
+          dirty: true,
+          status: 'INVALID',
+          value: { name: 'Ozan Turhan' },
+          path: 'student.studentForm'
+        }),
+        new ResetForm({ path: 'student.studentForm', value: { name: 'Mehmet Ozan Turhan' } })
+      ])
+      .toPromise();
+
+    const form = store.selectSnapshot(StudentState.getStudentForm);
+
+    // Assert
+    expect(form).toEqual({ model: { name: 'Mehmet Ozan Turhan' } });
   });
 
   describe('NgxsFormPlugin runtime behavior', () => {
@@ -604,7 +651,6 @@ describe('NgxsFormPlugin', () => {
         });
       }
 
-      // Act
       TestBed.configureTestingModule({
         imports: [
           ReactiveFormsModule,
@@ -618,6 +664,7 @@ describe('NgxsFormPlugin', () => {
       const fixture = TestBed.createComponent(MockComponent);
       fixture.detectChanges();
 
+      // Act
       const statuses: string[] = [];
 
       const subscription = actions$
@@ -644,6 +691,204 @@ describe('NgxsFormPlugin', () => {
 
       // Assert
       expect(statuses).toEqual(['INVALID', 'VALID']);
+    });
+
+    it('should reset form when ResetForm action call', () => {
+      @State({
+        name: 'todos',
+        defaults: {
+          todosForm: {
+            model: {
+              text: 'Buy some coffee'
+            },
+            dirty: true
+          }
+        }
+      })
+      @Injectable()
+      class TodosState {}
+
+      @Component({
+        template: `
+          <form [formGroup]="form" ngxsForm="todos.todosForm">
+            <input formControlName="text" /> <button type="submit">Add todo</button>
+          </form>
+        `
+      })
+      class MockComponent {
+        public form = new FormGroup({
+          text: new FormControl()
+        });
+      }
+
+      TestBed.configureTestingModule({
+        imports: [
+          ReactiveFormsModule,
+          NgxsModule.forRoot([TodosState]),
+          NgxsFormPluginModule.forRoot()
+        ],
+        declarations: [MockComponent]
+      });
+
+      const store = getStore();
+      const fixture = TestBed.createComponent(MockComponent);
+
+      fixture.detectChanges();
+
+      // Act
+      store.dispatch(new ResetForm({ path: 'todos.todosForm' }));
+
+      fixture.detectChanges();
+
+      // Arrange
+      const input = fixture.debugElement.query(By.css('form input')).nativeElement;
+
+      // Assert
+      expect(store.selectSnapshot(({ todos }) => todos).todosForm).toEqual({
+        dirty: false,
+        errors: {},
+        model: {
+          text: null
+        },
+        status: 'VALID'
+      });
+      expect(input.value).toEqual('');
+    });
+
+    it('should reset form when ResetForm action call after input changes', fakeAsync(() => {
+      // Arrange
+      @State({
+        name: 'todos',
+        defaults: {
+          todosForm: {
+            model: {
+              text: 'Buy some coffee'
+            },
+            dirty: true
+          }
+        }
+      })
+      @Injectable()
+      class TodosState {}
+
+      @Component({
+        template: `
+          <form [formGroup]="form" [ngxsFormDebounce]="500" ngxsForm="todos.todosForm">
+            <input formControlName="text" /> <button type="submit">Add todo</button>
+          </form>
+        `
+      })
+      class MockComponent {
+        public form = new FormGroup({
+          text: new FormControl()
+        });
+      }
+
+      TestBed.configureTestingModule({
+        imports: [
+          ReactiveFormsModule,
+          NgxsModule.forRoot([TodosState]),
+          NgxsFormPluginModule.forRoot()
+        ],
+        declarations: [MockComponent]
+      });
+
+      const store = getStore();
+      const fixture = TestBed.createComponent(MockComponent);
+      fixture.detectChanges();
+
+      const input = fixture.debugElement.query(By.css('form input')).nativeElement;
+      input.value = 'Buy some tea';
+      input.dispatchEvent(new KeyboardEvent('input'));
+
+      tick(200);
+
+      expect(store.selectSnapshot(({ todos }) => todos).todosForm).toEqual({
+        dirty: true,
+        model: {
+          text: 'Buy some coffee'
+        },
+        status: 'VALID'
+      });
+
+      // Act
+      store.dispatch(new ResetForm({ path: 'todos.todosForm' }));
+
+      tick(500);
+
+      // Assert
+      expect(store.selectSnapshot(({ todos }) => todos).todosForm).toEqual({
+        dirty: false,
+        errors: {},
+        model: {
+          text: null
+        },
+        status: 'VALID'
+      });
+    }));
+
+    it('should reset form with value when ResetForm action call', () => {
+      @State({
+        name: 'todos',
+        defaults: {
+          todosForm: {
+            model: {
+              text: 'Buy some coffee'
+            },
+            dirty: true
+          }
+        }
+      })
+      @Injectable()
+      class TodosState {}
+
+      @Component({
+        template: `
+          <form [formGroup]="form" ngxsForm="todos.todosForm">
+            <input formControlName="text" /> <button type="submit">Add todo</button>
+          </form>
+        `
+      })
+      class MockComponent {
+        public form = new FormGroup({
+          text: new FormControl()
+        });
+      }
+
+      TestBed.configureTestingModule({
+        imports: [
+          ReactiveFormsModule,
+          NgxsModule.forRoot([TodosState]),
+          NgxsFormPluginModule.forRoot()
+        ],
+        declarations: [MockComponent]
+      });
+
+      const store = getStore();
+      const fixture = TestBed.createComponent(MockComponent);
+
+      fixture.detectChanges();
+
+      // Act
+      store.dispatch(
+        new ResetForm({ path: 'todos.todosForm', value: { text: 'Buy some tea' } })
+      );
+
+      fixture.detectChanges();
+
+      // Arrange
+      const input = fixture.debugElement.query(By.css('form input')).nativeElement;
+
+      // Assert
+      expect(store.selectSnapshot(({ todos }) => todos).todosForm).toEqual({
+        dirty: false,
+        errors: {},
+        model: {
+          text: 'Buy some tea'
+        },
+        status: 'VALID'
+      });
+      expect(input.value).toEqual('Buy some tea');
     });
   });
 });